requirements.txt
numpy==1.19.2
pandas==1.1.2
matplotlib==3.3.2
torch==1.6.0
pyyaml==5.3.1